split argument handling methods into a separate module
