+++ lib/adafruit/io/client.rb
@@ -5,6 +5,7 @@ require 'faraday'
 require 'faraday_middleware'
 require 'faraday_middleware/response/mashify'
 
+require 'adafruit/io/arguments'
 require 'adafruit/io/configurable'
 require 'adafruit/io/request_handler'
 
@@ -17,6 +18,7 @@ module Adafruit
   module IO
     class Client
 
+      include Adafruit::IO::Arguments
       include Adafruit::IO::Configurable
       include Adafruit::IO::RequestHandler
 
@@ -66,49 +68,6 @@ module Adafruit
           c.use :instrumentation
           c.adapter Faraday.default_adapter
         end
-      end
-
-      # Allows us to give a username during client initialization or with a specific method.
-      def extract_username(args)
-        username = @username
-
-        if args.last.is_a?(Hash) && args.last.has_key?(:username)
-          username = args.last.delete(:username)
-        end
-
-        if username.nil?
-          raise "cannot make request when username is nil"
-        end
-
-        [ username, args ]
-      end
-
-      def get_key_from_arguments(arguments)
-        record_or_key = arguments.shift
-        return nil if record_or_key.nil?
-
-        if record_or_key.is_a?(String)
-          record_or_key
-        elsif record_or_key.is_a?(Hash) && record_or_key.has_key?('key')
-          record_or_key['key']
-        elsif record_or_key.is_a?(Hash) && record_or_key.has_key?(:key)
-          record_or_key[:key]
-        elsif record_or_key.respond_to?(:key)
-          record_or_key.key
-        else
-          raise 'unrecognized object or key value in arguments'
-        end
-      end
-
-      def get_query_from_arguments(arguments, params)
-        query = {}
-        options = arguments.shift
-        return query if options.nil?
-
-        params.each do |param|
-          query[param] = options[param.to_sym] if options.has_key?(param.to_sym)
-        end
-        query
       end
 
       def api_url(username, *args)
